requirements.txt
decorator==4.0.2
Django==1.8
django-extensions==1.5.5
django-picklefield==0.3.1
djangorestframework==3.2.2
extras==0.0.3
fuzzywuzzy==0.6.1
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
jsonfield==1.0.3
lettuce==0.2.20
linecache2==1.0.0
mock==1.0.1
path.py==7.6
pbr==1.5.0
pexpect==3.3
pickleshare==0.5
python-mimeparse==0.1.4
python-subunit==1.1.0
simplegeneric==0.8.1
six==1.9.0
sure==1.2.13
testtools==1.8.0
traceback2==1.4.0
traitlets==4.0.0
unittest2==1.1.0
